+++ packages/web-components/src/components/global-header/components/global-header/src/components/CommonHeader/CommonHeader.ts
@@ -25,6 +25,8 @@ import './WideSideNav';
 import '../SideNavItem/SideNavItem';
 import '../HeaderContext/HeaderContext';
 import useScript from '../../globals/useScript';
+import loadSidekickScript from '../../globals/loadSidekickScript';
+import loadSolisScript from '../../globals/loadSolisScript';
 
 const { stablePrefix: clabsPrefix } = settings;
 
@@ -51,6 +53,12 @@ export class CommonHeader extends LitElement {
   @state()
   assistMeScriptLoaded = false;
 
+  @state()
+  sidekickScriptLoaded = false;
+
+  @state()
+  solisScriptLoaded = false;
+
   handleNavItemClick = (e: Event) => {
     if (this.headerProps?.sideNav?.onClick) {
       this.headerProps?.sideNav?.onClick?.(e);
@@ -74,10 +82,50 @@ export class CommonHeader extends LitElement {
         }
       }) as EventListener);
     }
+    if (
+      this.headerProps.sidekickConfig &&
+      this.headerProps.sidekickConfig.isEnabled
+    ) {
+      loadSidekickScript(this.headerProps);
+
+      document.addEventListener('sidekick-script-status', ((e: CustomEvent) => {
+        if (
+          e.detail?.message &&
+          e.detail?.message === 'load' &&
+          this.headerProps.sidekickConfig
+        ) {
+          this.sidekickScriptLoaded = true;
+        } else if (e.detail?.message && e.detail?.message === 'error') {
+          console.error(
+            'An error occurred trying to load the sidekick script.'
+          );
+        }
+      }) as EventListener);
+    }
+    if (
+      this.headerProps.solisConfig &&
+      this.headerProps.solisConfig.isEnabled
+    ) {
+      loadSolisScript(this.headerProps);
+
+      document.addEventListener('solis-script-status', ((e: CustomEvent) => {
+        if (
+          e.detail?.message &&
+          e.detail?.message === 'load' &&
+          this.headerProps.solisConfig
+        ) {
+          this.solisScriptLoaded = true;
+        } else if (e.detail?.message && e.detail?.message === 'error') {
+          console.error('An error occurred trying to load the solis script.');
+        }
+      }) as EventListener);
+    }
   }
 
   disconnectedCallback() {
     this.removeEventListener('assist-me-script-status', () => {});
+    this.removeEventListener('sidekick-script-status', () => {});
+    this.removeEventListener('solis-script-status', () => {});
     super.disconnectedCallback();
   }
 
@@ -87,6 +135,18 @@ export class CommonHeader extends LitElement {
       changedProperties.has('headerProps')
     ) {
       useScript(this.headerProps);
+    }
+    if (
+      this.headerProps.sidekickConfig?.isEnabled &&
+      changedProperties.has('headerProps')
+    ) {
+      loadSidekickScript(this.headerProps);
+    }
+    if (
+      this.headerProps.solisConfig?.isEnabled &&
+      changedProperties.has('headerProps')
+    ) {
+      loadSolisScript(this.headerProps);
     }
   }
 
@@ -109,15 +169,15 @@ export class CommonHeader extends LitElement {
         <span class="${AUTOMATION_NAMESPACE_PREFIX}__capability-name"
           >${this.headerProps?.capabilityName?.label ?? nothing}</span
         >
-        <apaas-header-context
+        <clabs-global-header-context
           class="${AUTOMATION_NAMESPACE_PREFIX}__global"
           .props="${{ ...this.headerProps }}"
           .assistMeScriptLoaded="${this.assistMeScriptLoaded}"
           ?hasNewNotifications="${this
-            .hasNewNotifications}"></apaas-header-context>
+            .hasNewNotifications}"></clabs-global-header-context>
         ${this.headerProps && this.headerProps?.sideNav
           ? html`
-              <apaas-wide-side-nav
+              <clabs-global-header-wide-side-nav
                 aria-label=${this.headerProps?.sideNav?.buttonLabel ??
                 'Side navigation'}
                 collapse-mode="${typeof this.headerProps.sideNav
@@ -147,7 +207,7 @@ export class CommonHeader extends LitElement {
                               ${group?.links?.map((link, key, { length }) => {
                                 // Loop through the links array to render the menu items
                                 return html`
-                                  <apaas-side-nav-item
+                                  <clabs-global-header-side-nav-item
                                     .link="${{ ...link }}"
                                     .isCollapsible="${this.headerProps.sideNav
                                       ?.isCollapsible}"
@@ -164,7 +224,7 @@ export class CommonHeader extends LitElement {
                                     'function'}">
                                     .isHybridIpaas="${this.headerProps
                                       .isHybridIpaas}"
-                                  </apaas-side-nav-item>
+                                  </clabs-global-header-side-nav-item>
                                   ${
                                     // do not add a divider after the final group
                                     index + 1 !== numberOfGroups &&
@@ -187,7 +247,7 @@ export class CommonHeader extends LitElement {
                         ${this.headerProps.sideNav?.links?.map((link) => {
                           // Loop through the links array to render the menu items
                           return html`
-												<apaas-side-nav-item 
+												<clabs-global-header-side-nav-item
 												.link="${{ ...link }}"
 												.isCollapsible="${this.headerProps.sideNav?.isCollapsible}"
 												.handleNavItemClick="${this.handleNavItemClick}"
@@ -197,15 +257,15 @@ export class CommonHeader extends LitElement {
 												.isOnClickAvailable="${
                           typeof this.headerProps?.sideNav?.onClick ===
                           'function'
-                        }">	
-												</apaas-side-nav-item>
+                        }">
+												</clabs-global-header-side-nav-item>
                                 </cds-custom-side-nav-menu>
                               `;
                         })}
                       `
                     : nothing}
                 </cds-custom-side-nav-items>
-              </apaas-wide-side-nav>
+              </clabs-global-header-wide-side-nav>
             `
           : nothing}
         ${this.headerProps && this.headerProps.sideNavPropsV2
@@ -233,7 +293,7 @@ export class CommonHeader extends LitElement {
               ${this.headerProps.sideNavPropsV2?.links?.map((link) => {
                 // Loop through the links array to render the menu items
                 return html`
-									<apaas-side-nav-item 
+									<clabs-global-header-side-nav-item
 									.link="${{ ...link }}"
 									.isCollapsible="${this.headerProps.sideNavPropsV2?.isCollapsible}"
 									.handleNavItemClick="${this.handleNavItemClick}"
@@ -245,7 +305,7 @@ export class CommonHeader extends LitElement {
                     'function'
                   }"
 									>
-									</apaas-side-nav-item>
+									</clabs-global-header-side-nav-item>
                 </cds-custom-side-nav-menu>
                 `;
               })}