+++ packages/utilities/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.18.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/utilities@0.17.0...@carbon-labs/utilities@0.18.0) (2025-12-01)
+
+
+### Bug Fixes
+
+* utilities exports ([#927](https://github.com/carbon-design-system/carbon-labs/issues/927)) ([e0a636d](https://github.com/carbon-design-system/carbon-labs/commit/e0a636da31cf30c251be0cfe3ed79bb8f9fa086e))
+
+
+
+
+
 # [0.17.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/utilities@0.16.0...@carbon-labs/utilities@0.17.0) (2025-07-09)
 
 **Note:** Version bump only for package @carbon-labs/utilities